Do not print the AVF CSR to stdout
We will still validate the CSR, it just should not be printed to
stdout. The issue is that AVF CSRs do not have valid DeviceInfo, so
the RKP server will reject them.

Any device registered for KeyMint RKP is implicily registered to support
AVF as well.

There are tests in place to ensure the UDS_pub is the same for both
AVF and KeyMint. This means we can rely on the KeyMint CSR to reliably
register the device.

Bug: 421327176
Test: ran rkp_factory_extraction_tool on a pixel, verified output
Change-Id: I43dc0aed8fd8c8670acebbf93e8b193ad87ee43e