Split out km_compat tests
These tests only use a single `pub` entrypoint so can be an integration
test rather than a unit test.

Add a common function for ensuring that the compat service is started.

Also add some println!s to make it clear when/why tests are skipped,
and add a TEST_MAPPING file.

Note that this test only does anything on a device with a Keymaster
implementation; on a KeyMint device, all tests are skipped.

Bug: 379008955
Test: keystore2_km_compat_test
Flag: EXEMPT test refactor
Change-Id: I6fab2b147fcdfeadf0b3d9b02661d9153ec021eb