Avoid re-use of previous KeyType value
Early versions of `keystore2` included a `KeyType::Attestation` value as
an option (with value 2) in the `key_type` column of the `keyentry`
database table.

This was removed in aosp/2587525 when RKPD took over attestation key
management.   However, I don't _think_ there was any action taken to
clean up existing rows with `key_type==2`.

So add a comment to avoid the use of value 2 in any future extensions to
the `KeyType` enum.

Test: none, comment only
Flag: EXEMPT comment only
Change-Id: I8606815a40f125d8e735e9d626d4f818983ba1a3