Depend on _fake_latest target for C++ KeyMint
The `keystore2` service depends on the latest version of the KeyMint
HAL, but also includes a copy of the C++ KeyMint implementation code (to
support software-emulation of some features).  The C++ KeyMint code is
not going to support KeyMint versions above v4, so this will lead to
multiple-version errors when the KeyMint AIDL moves to v5.

To work around this, use the variant of the C++ target that claims to
support the latest KeyMint HAL, even though the code actually only
supports v4.

Test: TreeHugger
Flag: EXEMPT build refactor
Bug: 379008955
Change-Id: I47358f477645f49400aa4bf2e07be00e047568fa