Use oversampling for generating the prng seed
Previously, we read 48 bytes directly from the HWRNG to use as
the DRBG seed.

This patch implements an oversampling and conditioning step to
improve the quality and robustness of the initial seed and
reseeds. We now read 192 bytes (RAW_ENTROPY_SAMPLE_LEN) from the
HWRNG and hash this sample using SHA-512. The 64-byte digest
is then truncated to 48 bytes (drbg::ENTROPY_LEN) to be used
as the entropy input.

This conditioning step helps to diffuse the entropy from the
raw sample and protects against potential non-uniformity or
minor biases in the HWRNG output.

This change also introduces `libzeroize` to securely clear
sensitive entropy material (both the raw sample and the
conditioned seed) from memory. `libhex` is added as a
test-only dependency.

Bug: 444185751
Test: atest prng_seeder.test
Change-Id: I9e30e34c15c6bf0f8eb846d6e804af4daa568fe2