Retry generateKey with swapped IMEI tags on attestation ID mismatch.
Implements a workaround in `KeystoreSecurityLevel::generate_key` to
handle `CANNOT_ATTEST_IDS` errors from KeyMint when IMEI attestation is
requested. If such an error occurs, and IMEI tags are present, the
function retries the `generateKey` call with `ATTESTATION_ID_IMEI` and
`ATTESTATION_ID_SECOND_IMEI` tags swapped. This addresses issues on
older KeyMint implementations that may enforce stable ordering of IMEI
attestation.

This patch does not add support for doing this with imported keys or
wrapped imported keys. Support for those will need to be added in a
follow-up patch.

Bug: 437139792
Test: atest keystore2_client_tests

Change-Id: If829e40f6e65853c05424a62d9c62c8df39b7554