Explicitly pass key origin for metrics
The key creation metrics are based on the parameters passed in for a key
creation event; however, the KeyOrigin for a key is only included in the
key characteristics retrieved after the creation.

So explicitly pass the origin value into the metrics code (and assume
that the KeyMint implementation will correctly populate Origin based
on the creation method invoked).

Bug: 445143423
Test: check contents of dumpsys android.security.maintenance
Test: keystore2_test -- metrics_store
Flag: EXEMPT bug fix
Change-Id: Ic1550d0ab4643c7806fe21a44579be4cf4bf2bf0