Add common TA traits ConstTimeEq and MonotonicClock
These traits are common across TA implementations that back HALs.

MonotonicClock returns the MillisecondsSinceEpoch type. This type is
used as a wire type for some HALs so it's added in a new module called
types in the hal-wire crate.

AsCborValue is implemented manually for MillisecondsSinceEpoch because
in order to support deriving AsCborValue, the hal-derive macro would
need to be modified to refer to crate::AsCborValue rather than
hal_wire::AsCborValue. Such a change would then require the local
crate where the macro is used to import AsCborValue, CborError and
cbor_type_error, which is not preferable.

Bug: 432285494
Test: cargo build
Change-Id: Iea4576e6850478224a63a4d4f7cbeb20ff1369f2