Handle auth notifications in background thread
For authorization notifications that have no return value, we should not
hold up the calling process awaiting completion of processing.  So pass
the notifications into a queue, and have a separate background thread
that services that queue.  (This is roughly equivalent to making these
particular methods `oneway`.)

Also include a way for checks on the current lock status to be
synchronized with the state of the queue, and use this when checking
the device lock status (in `Enforcements::is_device_locked`).

Bug: 421860260
Bug: 462372417
Test: keystore2_test keystore2_client_tests
Test: locally add delays to lock/unlock processing and check
      Sync is only used when queue is nonempty
Flag: android.security.keystore2.async_lock_state
FIXED_READ_ONLY_FLAG_REASON=NATIVE_FLAG_PRE_METADATA_MOUNT

Change-Id: I69ef168d146484cba9d4e1af2f2d04ba89a00dc0