Check auth-type/SIDs in auth-per-op HATs
We've observed at least one failure in the use of an auth-per-op key
because the SIDs in the op-specific HAT don't match the key.

Enhance keystore2 to emit an error log for this situation, by keeping
track of the auth-type and SIDs required for an auth-per-op operation.

Bug: 447129026
Test: keystore2_client_tests
Flag: EXEMPT BUGFIX
Change-Id: I58688529bb7230814f9a1f3bfb39133c1d0bed66