Mark libkeystore-engine as Apex Available.
Historically the wpa_supplicant binary imported
a special variant of this library called
libkeystore-engine-wifi-hidl. Importantly,
that variant is only available to binaries in
the vendor partition.

As we move the binary from the vendor partition
to the mainline module, we will need to
switch to the non-vendor variant of this library.
This requires that library is be marked as
Apex Available for use in the mainline module.

Some Keystore libraries have already been made
Apex Available for our project, for instance
in ag/35428948.

Bug: 468372035
Test: Import the library in the mainline supplicant
      and build using m
Change-Id: I53a7c9eabe3841fd8e81ac22067a9c2e91316258