refactor(rkp-factory-extraction-tool): Rename RKPVM_INSTANCE_NAME
This change renames all instances of RKPVM_INSTANCE_NAME to
AVF_INSTANCE_NAME in rkp_factory_extraction_tool.cpp to align with
the change in ag/37579455. The term "RKP VM" is deprecated and this
change replaces it with the more accurate and up-to-date term "AVF"
to avoid tying the tool to implementation details.

Bug: 463592774
Test: Build and run rkp_factory_extraction_tool
Change-Id: Ic891e6611d5fd9179f2d9cc26cf99e5c7204d300