refactor(rkp-factory-extraction-tool): Rename RKPVM_INSTANCE_NAME
This change renames all instances of RKPVM_INSTANCE_NAME to
AVF_INSTANCE_NAME in rkp_factory_extraction_tool.cpp to align with
the change in ag/37579455. The term "RKP VM" is deprecated and this
change replaces it with the more accurate and up-to-date term "AVF"
to avoid tying the tool to implementation details.

Bug: 463592774
Test: Build and run rkp_factory_extraction_tool
Change-Id: I3dc131d947206c020280c7f5e263f31ea6e3d716